feat: ansible setup for building ei|docs|store|pms
I've chosen to have a separate role for each service; despite the fact that a
single one with a loop would be much easier to write right now. As the services
evolve with time, pretty soon the need arise to customize building each service
individually. When that happens, a single role often cause things to go haywire.