feat: add api-prefix support in pms app
This was the quickest way to get pms docs work when the service is hosted on a
sub-path. Unfortunate consequence of this is that when the app is hosted on a
sub-path as well as exposed on a port of its own; the api-prefix is going to
stay. i.e when the service is exposed on port 7002 of lab.channikhabra.com, to
host the service on lab.channikhabra.com/hot/pms, lab.channikhabra.com:7002/api
actually becomes lab.channikhabra.com:7002/pms/api