feat: database support with migrations
using postgresql-simple-migration for migrations.

I tried to use golang-migrate tool that we are using for store here as well, but
its god-awful API is just too much pain to work with. I am not sure how good or
bad postgresql-simple-migration is going to be, but the initial setup was a
breeze. A major win is that migrations are run as a part of app startup itself,
so separate ansible steps won't be required. It also act as a database
connection check, as well as a sanity check for schema health. It don't make
sense to run the service with invalid database schema.