feat: persist authenticated user across refreshes
Using redux-presist, but only for persisting authenticated user. I tried to
persist everything, but we are making the assumption in a few places that dates
in state are going to be day.js objects. Rehyderating from redux-persist breaks
the app because of this. I am thinking about making a few changes in the app's
architecture, which will take this concept (of having class instances in state)
further. Because of this, I am limiting the use of redux-persist.