feat: add standup SQL view
Better keep it in one place so anyone who need to read standups can do so
without copying over the query. It will also ease the transition when we decide
to create different tables/databases to keep materialized state for events.
Performing JSON queries on a gigantic table which is supposed to be a
append-only log will not scale too far.